Complete Refactor of Design
There is now a queue that is gathering the tasks and work one by one.

The queue will be priorized with the prio set in task

The call to execute a task will give back a uuid with this id it should
be possible to retreive information for current state of the given task.